Move spotify service into a sub-folder
