+++ apps/api/src/app/modules/spotify/spotify/spotify.service.spec.ts
